tripplanner/requirements.txt
flask
pymongo
openai
requests
python-dotenv
